feat(components/App and client/app): update app file to render the Welcome & Game View
