+++ client/app.js
@@ -1,33 +1,33 @@
-import React from 'react';
+// import React from 'react';
 
-import {Navbar} from './components';
-import Routes from './routes';
-import {makeStyles} from '@material-ui/core/styles';
-import CssBaseline from '@material-ui/core/CssBaseline';
-import Footer from './components/Footer';
+// import {Navbar} from './components';
+// import Routes from './routes';
+// import {makeStyles} from '@material-ui/core/styles';
+// import CssBaseline from '@material-ui/core/CssBaseline';
+// import Footer from './components/Footer';
 
-const useStyles = makeStyles(theme => ({
-  app: {
-    display: 'flex',
-    flexDirection: 'column',
-    minHeight: '100vh'
-  },
-  main: {
-    marginTop: theme.spacing(8),
-    marginBottom: theme.spacing(2)
-  }
-}));
+// const useStyles = makeStyles(theme => ({
+//   app: {
+//     display: 'flex',
+//     flexDirection: 'column',
+//     minHeight: '100vh'
+//   },
+//   main: {
+//     marginTop: theme.spacing(8),
+//     marginBottom: theme.spacing(2)
+//   }
+// }));
 
-const App = () => {
-  const classes = useStyles();
-  return (
-    <div id="app" className={classes.app}>
-      <CssBaseline />
-      <Navbar />
-      <Routes className={classes.main} />
-      <Footer />
-    </div>
-  );
-};
+// const App = () => {
+//   const classes = useStyles();
+//   return (
+//     <div id="app" className={classes.app}>
+//       <CssBaseline />
+//       <Navbar />
+//       {/* <Routes className={classes.main} /> */}
+//       <Footer />
+//     </div>
+//   );
+// };
 
-export default App;
+// export default App;
